migrated mission module queue funcs to new files
